Add local model support and Apple Silicon optimizations
- Update model loading to prioritize local files over HuggingFace downloads
- Add local model directory structure support (./models/FLUX.1-Krea-dev/)
- Fix Apple Silicon MPS compatibility issues (float64 → float32)
- Create Apple Silicon optimized inference script (flux_apple_silicon.py)
- Add Diffusers pipeline integration (flux_diffusers.py)
- Fix device handling and memory management for MPS
- Add model download utility and test scripts
- Successfully generate images on Apple Silicon with local models

+++ src/flux/math.py
@@ -13,7 +13,7 @@ def rope(pos: Tensor, dim: int, theta: int) -> Tensor:
     assert dim % 2 == 0
 
     b, l = pos.shape
-    scale = torch.arange(0, dim, 2, dtype=torch.float64, device=pos.device) / dim
+    scale = torch.arange(0, dim, 2, dtype=torch.float32, device=pos.device) / dim
     omega = 1.0 / ((theta) ** scale)
     
     out = torch.einsum("...n,d->...nd", pos, omega)
